Title: 낮은 지점들, Time: 116ms, Memory: 2MB, Status: Passed - CodeTree

+++ 240507/낮은 지점들/lowest-points.cpp
@@ -6,7 +6,7 @@ int main()
 {
     unordered_map<int, int> umap;
     int n;
-    int max = 0;
+    long max = 0;
     cin >> n;
     while(n--)
     {
